safari-534.53-branch/LayoutTests/fast/parser/block-nesting-cap-table-expected.txt

Test that the HTML parser does not allow the nesting depth of "block-level" elements to exceed 4096 when using nested table tag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL t1024.parentNode === td1022 should be true. Was false.
PASS t1023.parentNode === td1022 is true
FAIL t1024.previousSibling === t1023 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
